backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
# readme = "README.md"  # Removed because file doesn't exist
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "gunicorn>=23.0.0",
    "rest-framework-simplejwt>=0.0.2",
    "djangorestframework-simplejwt>=5.5.1",
    "redis>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.12.0",
    "django-extensions>=4.1",
    "ipython>=9.8.0",
    "isort>=7.0.0",
    "pycodestyle>=2.14.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
